Fix case-sensitivity in choosing output filenames.
The URL format of original page images seems to have changed somewhat,
and the issue ID in it is now upper-case, as it is in other API fields,
as opposed to how it is in end-user-facing URLs.

This is something of a workaround; the more robust solution would be to
determine the issue ID from the page ID in the API response, and not
take an `issue` argument to `fetch_pages()` at all. I want to revisit
this as part of making it possible to download single pages or a range
of pages, not just the whole issue.